% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siggaard_andersen_co2_params.R
\name{actual_bicarbonate_content_mmol_l}
\alias{actual_bicarbonate_content_mmol_l}
\title{Calculate plasma bicarbonate concentration as per Henderson-Hasselbalch equation.}
\usage{
actual_bicarbonate_content_mmol_l(
  pco2,
  ph = 7.4,
  pco2_units = c("kPa", "mmHg")
)
}
\arguments{
\item{pco2}{CO2 partial pressure}

\item{ph}{pH (hydrogen ion concentration). Default 7.40}

\item{pressure_units}{Unit for \code{pco2}; one of \code{"kPa"} or \code{"mmHg"}.}
}
\value{
The HCO3 concentration of plasma in mmol/dL
}
\description{
\code{actual_bicarbonate_content_mmol_l} calculates plasma bicarbonate concentration as per the
Henderson-Hasselbalch equation.
}
\details{
Calculation is a straightforward Henderson-Hasselbalch rearrangement. This relies
on the solubility coefficient of CO2 in plasma. \insertCite{siggaard_1988}{co2ntent}
provide a constant of 0.230 mmol/L/kPa.
}
\references{
{
  \insertRef{douglas_1988}{co2ntent}
  \insertRef{siggaard_1988}{co2ntent}
}
}
\keyword{internal}
