% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serum_solubility_pk.R
\name{apparent_pk}
\alias{apparent_pk}
\title{Calculate the apparent pK' of the CO2-bicarbonate system in blood.}
\usage{
apparent_pk(t, ph)
}
\arguments{
\item{t}{Plasma Temperature.}

\item{ph}{Plasma pH.}
}
\value{
The apparent pK'
}
\description{
Calculate the apparent pK' of the CO2-bicarbonate system in blood.
}
