% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serum_solubility_pk.R
\name{co2_plasma_solubility}
\alias{co2_plasma_solubility}
\title{Calculate the solubility coefficient of CO2 in plasma. Austin et al(1963)}
\usage{
co2_plasma_solubility(t)
}
\arguments{
\item{t}{Plasma Temperature.}
}
\value{
The solubility coefficient of CO2 in plasma
}
\description{
Calculate the solubility coefficient of CO2 in plasma. Austin et al(1963)
}
