% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/molar_volumes.R
\name{.molar_volume_defaults}
\alias{.molar_volume_defaults}
\title{Default molar volumes for gases (internal)}
\usage{
.molar_volume_defaults()
}
\value{
Named numeric vector (dL/mmol).
}
\description{
Named numeric vector of default molar volumes in decilitres per millimole (dL/mmol).
Values are chosen for common physiological gases:
- co2: carbon dioxide
- o2: oxygen
- ideal: ideal gas approximation
}
\details{
These defaults are used by mmols_l_to_mls_dl() and mls_dl_to_mmols_l()
when `molar_volume` is not supplied.
}
\examples{
.molar_volume_defaults()["o2"]
}
\seealso{
mmols_l_to_mls_dl, mls_dl_to_mmols_l
}
\keyword{internal}
