% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/douglas_apparent_pk.R
\name{douglas_apparent_pk_co2_hco3}
\alias{douglas_apparent_pk_co2_hco3}
\title{Calculate the apparent pK' of the CO2-Bicarbonate system in blood.}
\usage{
douglas_apparent_pk_co2_hco3(
  temperature = 37,
  ph = 7.4,
  skip_range_check = FALSE
)
}
\arguments{
\item{temperature}{Plasma Temperature in Celsius. Default 37c}

\item{ph}{Plasma pH. Default 7.40}

\item{skip_range_check}{If TRUE skip checking of parameter ranges. Default: FALSE}
}
\value{
The apparent pK'
}
\description{
\code{douglas_apparent_pk_co2_hco3} calculates the apparent pK'of the
CO2-Bicarbonate in blood via the formula described by 
\insertCite{douglas_1988}{co2ntent}.
}
\references{
{
  \insertRef{douglas_1988}{co2ntent}
}
}
