% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/douglas_co2_plasma_solubility.R
\name{douglas_co2_plasma_solubility}
\alias{douglas_co2_plasma_solubility}
\title{Calculate the solubility coefficient of CO2 in plasma.}
\usage{
douglas_co2_plasma_solubility(temperature = 37, skip_range_check = FALSE)
}
\arguments{
\item{temperature}{Plasma Temperature in Celsius. Default 37c}

\item{skip_range_check}{If TRUE skip checking of parameter ranges. Default: FALSE}
}
\value{
s The solubility coefficient of CO2 in plasma mmol/dL/kPa
}
\description{
\code{douglas_co2_plasma_solubility} calculates the solubility coefficient
of CO2 in plasma via the formula described by \insertCite{douglas_1988}{co2ntent}.
}
\references{
{
  \insertRef{douglas_1988}{co2ntent}
}
}
