% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_checks.R
\name{haemoglobin_g_dl_param_check}
\alias{haemoglobin_g_dl_param_check}
\title{Check vector of Haemoglobin values.}
\usage{
haemoglobin_g_dl_param_check(haemoglobin_g_dl)
}
\arguments{
\item{haemoglobin_g_dl}{Haemoglobin g/dL.}
}
\value{
No return value.
}
\description{
\code{haemoglobin_g_dl_param_check} checks a vector of Haemoglobin values for abnormal or missing values.
Abnormal values are designed to catch values entered in error rather than representing a 'normal'
physiological range. If abnormal values are seen, a warning is raised. If missing values are
found an error is raised. Abnormal values are defined as Hb < 2g/dL  or Hb > 20 g/dL
}
