% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kelman_po2_to_so2.R
\name{kelman_virtual_po2}
\alias{kelman_virtual_po2}
\title{Calculate virtual pO2.}
\usage{
kelman_virtual_po2(
  po2,
  pco2,
  temperature = 37,
  ph = 7.4,
  pressure_units = c("kPa", "mmHg")
)
}
\arguments{
\item{po2}{O2 partial pressure}

\item{pco2}{CO2 partial pressure. Default 5.332895kPa (40mmHg)}

\item{temperature}{temperature in celsius. Default 37c}

\item{ph}{pH (hydrogen ion concentration). Default 7.40}

\item{pressure_units}{Input parameters are kPa, otherwise use mmHg}
}
\value{
Vector of virtual pO2
}
\description{
\code{kelman_virtual_po2} calculates a 'virtual po2', a pO2 corrected for
ph pco2 and temperature, as per the method described by
\insertCite{kelman_1966}{co2ntent}.
}
\details{
Units for pO2 and pCO2 should be in the same units (mmHg or kPa) as specified by pressure_units.
The result is returned in same units
}
\references{
{
  \insertRef{kelman_1966}{co2ntent}
}
}
\keyword{internal}
