% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loeppky_blood_co2_content_ml_dl.R
\name{loeppky_blood_co2_content_ml_dl}
\alias{loeppky_blood_co2_content_ml_dl}
\title{Calculate CO2 content of blood via Loeppky method}
\usage{
loeppky_blood_co2_content_ml_dl(pco2, pco2_units = c("kPa", "mmHg"))
}
\arguments{
\item{pco2}{CO2 partial pressure}

\item{pco2_units}{Unit for \code{pco2}; one of \code{"kPa"} or \code{"mmHg"}.}
}
\value{
The CO2 content of plasma in ml/dL
}
\description{
\code{loeppky_blood_co2_content_ml_dl} calculates the
CO2 content of blood via the method described by \insertCite{loeppky_1983}{co2ntent}.
}
\references{
{
  \insertRef{loeppky_1983}{co2ntent}
}
}
