% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/units_conversion.R
\name{mls_dl_to_mmols_dl}
\alias{mls_dl_to_mmols_dl}
\title{Convert mls/dL to mmols/dL}
\usage{
mls_dl_to_mmols_dl(mls_dl)
}
\arguments{
\item{mls_dl}{Content in mls/dL}
}
\value{
Content in mmols/dL
}
\description{
Convert mls/dL to mmols/dL
}
