% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/units_conversion.R
\name{mls_dl_to_mmols_l}
\alias{mls_dl_to_mmols_l}
\title{Convert mls/dL to mmols/L}
\usage{
mls_dl_to_mmols_l(mls_dl)
}
\arguments{
\item{mls_dl}{Content in mls/dL}
}
\value{
Content in mmols/L
}
\description{
Convert mls/dL to mmols/L
}
