% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/units_conversion.R
\name{mmols_dl_to_mls_dl}
\alias{mmols_dl_to_mls_dl}
\title{Convert mmols/dL to mls/dL}
\usage{
mmols_dl_to_mls_dl(mmols_dl)
}
\arguments{
\item{mmols_dl}{Content in mmols/dL}
}
\value{
Content in mls/dL
}
\description{
Convert mmols/dL to mls/dL
}
