% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/o2_content.R
\name{o2_content}
\alias{o2_content}
\title{High-level interface for O2 content of blood}
\usage{
o2_content(
  po2,
  so2_fraction,
  haemoglobin_g_dl,
  hufners_constant = 1.306,
  po2_units = c("kPa", "mmHg"),
  content_units = c("ml/dL", "mmol/L")
)
}
\arguments{
\item{po2}{O2 partial pressure.}

\item{so2_fraction}{Haemoglobin saturation as a fraction
(e.g. \eqn{0 < so2\_fraction < 1.0}).}

\item{haemoglobin_g_dl}{Haemoglobin concentration in g/dL.}

\item{hufners_constant}{Oxygen capacity of human haemoglobin in mL/g.
Default is \code{1.306} as per Gregory (1974).}

\item{po2_units}{Unit for \code{po2}; one of \code{"kPa"} or \code{"mmHg"}.}

\item{content_units}{Unit for result; one of \code{"mmol/L"} or \code{"ml/dL"}.}
}
\value{
Numeric vector of O2 content in mL/dL.
}
\description{
Provides a unified entry point to calculate the oxygen content of blood in
mL of O2 per dL of blood.
}
\examples{
o2_content(
  po2            = 10,
  so2_fraction   = 0.95,
  haemoglobin_g_dl = 15,
  po2_units = "kPa"
)

}
