% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_checks.R
\name{pco2_param_check}
\alias{pco2_param_check}
\title{Check vector of pCO2 values.}
\usage{
pco2_param_check(pco2, pco2_units = c("kPa", "mmHg"))
}
\arguments{
\item{pco2}{CO2 partial pressure}

\item{pco2_units}{Unit for \code{pco2}; one of \code{"kPa"} or \code{"mmHg"}.}
}
\value{
No return value.
}
\description{
\code{pco2_param_check} checks a vector of pCO2 values for abnormal or missing values.
Abnormal values are designed to catch values entered in error rather than representing a 'normal'
physiological range. If abnormal values are seen, a warning is raised. 
Abnormal values are defined as pCO2 < 1kPa or pCO2 > 20kPa
}
