% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siggaard_andersen_co2_params.R
\name{plasma_bicarbonate_content}
\alias{plasma_bicarbonate_content}
\title{Calculate plasma bicarbonate concentration as per Henderson-Hasselbach equation.}
\usage{
plasma_bicarbonate_content(pco2, ph = 7.4, temperature = 37,
  calculate_solubility_coefficient = FALSE, inputs_are_kpa = TRUE,
  skip_range_check = FALSE)
}
\arguments{
\item{pco2}{CO2 partial pressure}

\item{ph}{pH (hydrogen ion concentration). Default 7.40}

\item{temperature}{temperature in celcius. Default 37c}

\item{calculate_solubility_coefficient}{Default FALSE}

\item{inputs_are_kpa}{If TRUE, input pCO2 is in kPa, if FALSE use mmHg}

\item{skip_range_check}{If TRUE skip checking of parameter ranges. Default: FALSE}
}
\value{
The HCO3 concentration of plasma in mmol/dL
}
\description{
\code{plasma_bicarbonate_content} plasma bicarbonate concentration as per the
Henderson-Hasselbach equationv described by \insertCite{siggaard_1988}{co2ntent}.
}
\details{
Calculation is a straightforward Henderson-Hasselbach rearrangement. This relies
on the solubility coeffiecient of CO2 in plasma. \insertCite{siggaard_1988}{co2ntent}
provide a constant of 0.230 mmol/L/kPa, or we can calculate with the method provided
by \insertCite{douglas_1988}{co2ntent} implemented as \code{\link{co2_plasma_solubility}}
}
\references{
{
  \insertRef{douglas_1988}{co2ntent}
  \insertRef{siggaard_1988}{co2ntent}
}
}
