% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_checks.R
\name{po2_param_check}
\alias{po2_param_check}
\title{Check vector of pO2 values.}
\usage{
po2_param_check(po2, po2_units = c("kPa", "mmHg"))
}
\arguments{
\item{po2}{O2 partial pressure}

\item{po2_units}{Unit for \code{po2}; one of \code{"kPa"} or \code{"mmHg"}.}
}
\value{
No return value.
}
\description{
\code{po2_param_check} checks a vector of pO2 values for abnormal or missing values.
Abnormal values are designed to catch values entered in error rather than representing a 'normal'
physiological range. If abnormal values are seen, a warning is raised.
Abnormal values are defined as pO2 < 1kPa or pO2 > 100kPa
}
