% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/po2_to_so2.R
\name{po2_to_so2}
\alias{po2_to_so2}
\title{High-level interface for pO2 to SO2}
\usage{
po2_to_so2(
  po2,
  temperature = 37,
  ph = 7.4,
  pco2 = 5.332895,
  pressure_units = c("kPa", "mmHg"),
  method = c("kelman")
)
}
\arguments{
\item{po2}{O2 partial pressure.}

\item{temperature}{Temperature in Celsius. Default 37.}

\item{ph}{pH (hydrogen ion concentration). Default 7.40.}

\item{pco2}{CO2 partial pressure. Default 5.332895 kPa (40 mmHg).}

\item{pressure_units}{Units for \code{po2} and \code{pco2}; one of \code{"kPa"} or \code{"mmHg"}.}

\item{method}{Character string; currently only \code{"kelman"} is supported.}
}
\value{
Numeric vector of haemoglobin saturation as a fraction.
}
\description{
Calculates haemoglobin oxygen saturation from partial pressure of oxygen in
blood via a chosen method (currently Kelman only).
}
\examples{
po2_to_so2(
  po2  = 10,
  temperature = 37,
  ph   = 7.4,
  pco2 = 5.3329,
  pressure_units = "kPa"
)

}
