% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siggaard_andersen_co2_content.R
\name{siggaard_andersen_blood_co2_content_mmol_dl}
\alias{siggaard_andersen_blood_co2_content_mmol_dl}
\title{Calculate whole blood CO2 content as per Siggaard-Andersen method.}
\usage{
siggaard_andersen_blood_co2_content_mmol_dl(hco3_mmols_dl, pco2,
  haemoglobin_g_dl, so2_fraction, ph = 7.4, inputs_are_kpa = TRUE)
}
\arguments{
\item{hco3_mmols_dl}{plasma bicarbonate concentration mmols/dL}

\item{pco2}{CO2 partial pressure}

\item{haemoglobin_g_dl}{Haemoglobin g/dL. No default}

\item{so2_fraction}{Haemoglobin saturation as a fraction e.g 0 < so2_fraction < 1.0}

\item{ph}{pH (hydrogen ion concentration). Default 7.40}

\item{inputs_are_kpa}{If TRUE, input pCO2 is in kPa, if FALSE use mmHg}
}
\value{
The CO2 content of blood in mmol/dL
}
\description{
\code{siggaard_andersen_blood_co2_content_mmol_dl} calculates whole blood
CO2 content via the method described by \insertCite{siggaard_1988}{co2ntent}.
}
\references{
{
  \insertRef{siggaard_1988}{co2ntent}
}
}
