% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siggaard_andersen_co2_content.R
\name{siggaard_andersen_plasma_co2_content_mmol_l}
\alias{siggaard_andersen_plasma_co2_content_mmol_l}
\title{Calculate plasma CO2 content as per Siggaard-Andersen method.}
\usage{
siggaard_andersen_plasma_co2_content_mmol_l(
  hco3_mmols_l,
  pco2,
  pco2_units = c("kPa", "mmHg")
)
}
\arguments{
\item{hco3_mmols_l}{plasma bicarbonate concentration mmols/dL}

\item{pco2}{CO2 partial pressure}

\item{pco2_units}{Unit for \code{pco2}; one of \code{"kPa"} or \code{"mmHg"}}
}
\value{
The CO2 content of plasma in mmol/L
}
\description{
\code{siggaard_andersen_plasma_co2_content_mmol_l} calculates plasma CO2 content via the method described
by \insertCite{siggaard_1988}{co2ntent}.
}
\references{
{
  \insertRef{siggaard_1988}{co2ntent}
}
}
