% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_checks.R
\name{so2_fraction_param_check}
\alias{so2_fraction_param_check}
\title{Check vector of fractional SO2 values.}
\usage{
so2_fraction_param_check(so2_fraction, skip_range_check = FALSE)
}
\arguments{
\item{so2_fraction}{Haemoglobin saturation as a fraction}

\item{skip_range_check}{If TRUE skip checking of parameter ranges. Default: FALSE}
}
\value{
No return value.
}
\description{
\code{so2_fraction_param_check} checks a vector of fractional SO2 values for
abnormal or missing values. If abnormal values are seen, a warning is
raised. If missing values are found an error is raised. Abnormal values
are defined as so2_fraction < 0  or so2_fraction > 1
}
