% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_checks.R
\name{temperature_param_check}
\alias{temperature_param_check}
\title{Check vector of temperature values.}
\usage{
temperature_param_check(temperature)
}
\arguments{
\item{temperature}{temperature in celsius.}
}
\value{
No return value.
}
\description{
\code{temperature_param_check} checks a vector of temperature values for abnormal or missing values.
Abnormal values are designed to catch values entered in error rather than representing a 'normal'
physiological range. If abnormal values are seen, a warning is raised. If missing values are
found an error is raised. Abnormal values are defined as T < 28c  or T > 44c
}
