% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/co2ntent.R
\docType{package}
\name{co2ntent-package}
\alias{co2ntent}
\alias{co2ntent-package}
\title{co2ntent: A package for calculating carbon dioxide and oxygen content
of blood.}
\description{
The co2ntent package provides several formulae for the calculation of carbon
dioxide and oxygen content of the blood.
}
\details{
The co2ntent package provides several canonical formulae from the
literature, for the calculation of carbon dioxide and oxygen content
of the blood.

Data sets presented in the referenced literature are also included.

Several helper methods for units conversion and other minor tasks are
provided.
}
\section{co2ntent functions}{


\code{\link{douglas_apparent_pk_co2_hco3}}
\code{\link{douglas_blood_co2_content_ml_dl}}
\code{\link{douglas_co2_plasma_solubility}}
\code{\link{douglas_co2_plasma_to_blood_ratio}}
\code{\link{douglas_plasma_co2_content_ml_dl}}
\code{\link{kelman_po2_to_so2}}
\code{\link{kelman_virtual_po2}}
\code{\link{kelman_std_po2_to_so2}}
\code{\link{siggaard_andersen_erythrocyte_ph}}
\code{\link{siggaard_andersen_erythrocyte_p_k}}
\code{\link{siggaard_andersen_plasma_bicarbonate_content_mmol_l}}
\code{\link{siggaard_andersen_plasma_co2_content_mmol_l}}
\code{\link{siggaard_andersen_blood_co2_content_mmol_l}}
}

\author{
\strong{Maintainer}: Barnaby Akenzua-Sanderson \email{barnaby.sanderson1@nhs.net}

}
\keyword{internal}
