% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/douglas_co2_contents.R
\name{douglas_co2_plasma_to_blood_ratio}
\alias{douglas_co2_plasma_to_blood_ratio}
\title{Calculate the plasma to blood CO2 content ratio via Douglas method}
\usage{
douglas_co2_plasma_to_blood_ratio(
  haemoglobin_g_dl,
  so2_fraction,
  ph = 7.4,
  skip_range_check = FALSE
)
}
\arguments{
\item{haemoglobin_g_dl}{Haemoglobin g/dL. No default}

\item{so2_fraction}{Haemoglobin saturation as a fraction e.g 0 < so2_fraction < 1.0}

\item{ph}{pH (hydrogen ion concentration). Default 7.40}

\item{skip_range_check}{If TRUE skip checking of parameter ranges. Default: FALSE}
}
\value{
The CO2 content Plasma:Blood ratio
}
\description{
\code{douglas_co2_plasma_to_blood_ratio} calculates the ratio between the CO2
content of plasma and the CO2 content of whole blood via the method described by
\insertCite{douglas_1988}{co2ntent}.
}
\details{
The plasma/blood ratio is calculated via an equation of the form derived by
\insertCite{visser_1960}{co2ntent} and \insertCite{mchardy_1967}{co2ntent},
but with the coefficients derived and published by \insertCite{douglas_1988}{co2ntent}
}
\references{
{
  \insertRef{douglas_1988}{co2ntent}
  \insertRef{mchardy_1967}{co2ntent}
  \insertRef{visser_1960}{co2ntent}
}
}
