% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kelman_po2_to_so2.R
\name{kelman_std_po2_to_so2}
\alias{kelman_std_po2_to_so2}
\title{Calculate SO2 from pO2 of blood.}
\usage{
kelman_std_po2_to_so2(po2, inputs_are_kpa = TRUE, skip_range_check = FALSE)
}
\arguments{
\item{po2}{O2 partial pressure}

\item{inputs_are_kpa}{Input parameters are kPa, otherwise use mmHg}

\item{skip_range_check}{If TRUE skip checking of parameter ranges. Default: FALSE}
}
\value{
Haemoglobin saturation as fraction
}
\description{
\code{kelman_std_po2_to_so2} calculates haemoglobin oxygen saturation from partial pressure of oxygen in blood
viathe method described by \insertCite{kelman_1966}{co2ntent}. This form makes no correction for acid/base
disturbance, assuming pH=7.4, temperature=37c and pCO2=40mmHg
}
\note{
\code{kelman_std_po2_to_so2} doesn't work for po2 < 0.5kpa as returned SO2 can become negative
}
\references{
{
  \insertRef{kelman_1966}{co2ntent}
}
}
