% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kelman_po2_to_so2.R
\name{kelman_virtual_po2}
\alias{kelman_virtual_po2}
\title{Calculate virtual pO2.}
\usage{
kelman_virtual_po2(po2, pco2, temperature = 37, ph = 7.4,
  inputs_are_kpa = TRUE)
}
\arguments{
\item{po2}{O2 partial pressure}

\item{pco2}{CO2 partial pressure. Default 5.332895kPa (40mmHg)}

\item{temperature}{temperature in celcius. Default 37c}

\item{ph}{pH (hydrogen ion concentration). Default 7.40}

\item{inputs_are_kpa}{Input parameters are kPa, otherwise use mmHg}
}
\value{
Haemoglobin saturation as fraction
}
\description{
\code{kelman_virtual_po2} calculates a 'virtual po2', a pO2 corrected for ph pco2 and temperature, as
per the method described by \insertCite{kelman_1966}{co2ntent}.
}
\references{
{
  \insertRef{kelman_1966}{co2ntent}
}
}
