% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/units_conversion.R
\name{mmols_l_to_mls_dl}
\alias{mmols_l_to_mls_dl}
\title{Convert mmols/L to mls/dL}
\usage{
mmols_l_to_mls_dl(mmols_l)
}
\arguments{
\item{mmols_l}{Content in mmols/L}
}
\value{
Content in mls/dL
}
\description{
Convert mmols/L to mls/dL
}
