% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_checks.R
\name{ph_param_check}
\alias{ph_param_check}
\title{Check vector of ph values.}
\usage{
ph_param_check(ph)
}
\arguments{
\item{ph}{pH (hydrogen ion concentration)}
}
\value{
No return value.
}
\description{
\code{ph_param_check} checks a vector of pH values for abnormal or missing values.
Abnormal values are designed to catch values entered in error rather than representing a 'normal'
physiological range. If abnormal values are seen, a warning is raised. If missing values are
found an error is raised. Abnormal values are defined as pH < 6.8  or pH > 7.7
}
