% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siggaard_andersen_co2_content.R
\name{siggaard_andersen_blood_co2_content_mmol_l}
\alias{siggaard_andersen_blood_co2_content_mmol_l}
\title{Calculate whole blood CO2 content as per Siggaard-Andersen method.}
\usage{
siggaard_andersen_blood_co2_content_mmol_l(
  hco3_mmols_l,
  pco2,
  haemoglobin_g_dl,
  so2_fraction,
  ph = 7.4,
  inputs_are_kpa = TRUE,
  skip_range_check = FALSE
)
}
\arguments{
\item{hco3_mmols_l}{plasma bicarbonate concentration mmols/L}

\item{pco2}{CO2 partial pressure}

\item{haemoglobin_g_dl}{Haemoglobin g/dL. No default}

\item{so2_fraction}{Haemoglobin saturation as a fraction e.g 0 < so2_fraction < 1.0}

\item{ph}{pH (hydrogen ion concentration). Default 7.40}

\item{inputs_are_kpa}{If TRUE, input pCO2 is in kPa, if FALSE use mmHg}

\item{skip_range_check}{If TRUE skip checking of parameter ranges. Default: FALSE}
}
\value{
The CO2 content of blood in mmol/L
}
\description{
\code{siggaard_andersen_blood_co2_content_mmol_l} calculates whole blood
CO2 content via the method described by \insertCite{siggaard_1988}{co2ntent}.
}
\references{
{
  \insertRef{siggaard_1988}{co2ntent}
}
}
