% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siggaard_andersen_co2_params.R
\name{siggaard_andersen_plasma_bicarbonate_content_mmol_l}
\alias{siggaard_andersen_plasma_bicarbonate_content_mmol_l}
\title{Calculate plasma bicarbonate concentration as per Henderson-Hasselbalch equation.}
\usage{
siggaard_andersen_plasma_bicarbonate_content_mmol_l(
  pco2,
  ph = 7.4,
  temperature = 37,
  pco2_units = c("kPa", "mmHg")
)
}
\arguments{
\item{pco2}{CO2 partial pressure}

\item{ph}{pH (hydrogen ion concentration). Default 7.40}

\item{temperature}{temperature in celcius. Default 37c}

\item{pressure_units}{Unit for \code{pco2}; one of \code{"kPa"} or \code{"mmHg"}.}

\item{skip_range_check}{If TRUE skip checking of parameter ranges. Default: FALSE}
}
\value{
The HCO3 concentration of plasma in mmol/dL
}
\description{
\code{siggaard_andersen_plasma_bicarbonate_content_mmol_l} plasma bicarbonate concentration as per the
Henderson-Hasselbalch equation described by \insertCite{siggaard_1988}{co2ntent}.
}
\details{
Calculation is a straightforward Henderson-Hasselbalch rearrangement. This relies
on the solubility coefficient of CO2 in plasma. \insertCite{siggaard_1988}{co2ntent}
provide a constant of 0.230 mmol/L/kPa.
}
\references{
{
  \insertRef{douglas_1988}{co2ntent}
  \insertRef{siggaard_1988}{co2ntent}
}
}
\keyword{internal}
